requirements.txt
srt
pdfminer.six
pypandoc_binary
tqdm
google-cloud-texttospeech
